+++ BeerDispancer/BeerDispencer.Domain/Abstractions/IBeerFlowCalculator.cs
@@ -4,7 +4,7 @@ namespace BeerDispencer.Domain.Abstractions
 	public interface IBeerFlowCalculator
 	{
         public decimal? GetFlowVolume(DateTime? closedAt, DateTime? openAt);
-        public decimal? GetTotalSpent(double? volume);     
+        public decimal? GetTotalSpent(decimal? volume);     
     }
 }
