+++ BeerDispancer/BeerDispencer.Host/ViewModels/Response/Dispencer.cs
@@ -4,7 +4,7 @@ namespace BeerDispencer.WebApi.ViewModels.Response
 	public class Dispencer
     {
         public Guid Id { get; set; }
-        public double FlowVolume { get; set; }
+        public decimal FlowVolume { get; set; }
 		
     };
 }
